Moved new function to utils.
